Translate reg values before using them
Reg entries are bus local information. To get an address valid on
the "root bus" we have to traverse the tree and translate the "reg"
entries using "ranges" properties until we reach the root node.

Here we introduce structures and methods to handle cells and
structures composed using cells (reg, ranges, interrupts).

* cell.h
  * Cell
    * encapsulates data for one cell with up to 4 elements in it
    * actual size is specified via #address-cells, #size-cells,
      #interrupt-cells
    * assumes big endian
    * provides relational operators
    * provides add/sub operators
  * Range
    * provides methods to translate a cell from one bus domaon to another
  * Reg
    * encapsulates a reg entry

* device_tree.h
  * introduced translate_reg()
  * modified get_reg_val() to use translate_reg()

* modified users of get_reg_val() to use the new interface

+++ server/src/generic_guest.h
@@ -54,7 +54,7 @@ public:
   { _mmio_fallback = ds; }
 
   void register_mmio_device(cxx::Ref_ptr<Vmm::Mmio_device> &&dev,
-                            Vdev::Dt_node const &node, int index = 0);
+                            Vdev::Dt_node const &node, size_t index = 0);
 
   L4virtio::Ptr<void> load_ramdisk_at(char const *ram_disk,
                                       L4virtio::Ptr<void> addr,
